requirements.txt
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.10.6
tqdm==4.66.1
beautifulsoup4==4.12.3
html2text==2020.1.16
